Move permission object from shared to feature package
