+++ server/src/main/java/com/szampchat/server/permission/converters/PermissionToIntConverter.java
@@ -1,6 +1,6 @@
-package com.szampchat.server.shared.permission.converters;
+package com.szampchat.server.permission.converters;
 
-import com.szampchat.server.shared.permission.Permissions;
+import com.szampchat.server.permission.data.Permissions;
 import org.springframework.core.convert.converter.Converter;
 
 public class PermissionToIntConverter implements Converter<Permissions, Integer> {
