+++ server/src/main/java/com/szampchat/server/permission/repository/PermissionRepository.java
@@ -1,10 +1,6 @@
 package com.szampchat.server.permission.repository;
 
-import com.szampchat.server.shared.permission.PermissionOverrides;
-import org.springframework.data.r2dbc.repository.Query;
-import org.springframework.data.r2dbc.repository.R2dbcRepository;
-import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Repository;
+import com.szampchat.server.permission.data.PermissionOverrides;
 import reactor.core.publisher.Flux;
 
 //Not sure if it should be repo
