+++ server/src/main/java/com/szampchat/server/role/entity/Role.java
@@ -1,7 +1,6 @@
 package com.szampchat.server.role.entity;
 
-import com.szampchat.server.shared.permission.PermissionOverrides;
-import com.szampchat.server.shared.permission.Permissions;
+import com.szampchat.server.permission.data.PermissionOverrides;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
