+++ server/src/main/java/com/szampchat/server/role/repository/RoleRepository.java
@@ -1,7 +1,6 @@
 package com.szampchat.server.role.repository;
 
 import com.szampchat.server.role.entity.Role;
-import com.szampchat.server.shared.permission.Permissions;
 import org.springframework.data.r2dbc.repository.Query;
 import org.springframework.data.r2dbc.repository.R2dbcRepository;
 import org.springframework.data.repository.query.Param;
